refactor: split lower into convert parts of ast
